% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryOperations.R
\name{Buffer}
\alias{Buffer}
\title{Buffered point method}
\usage{
Buffer(pbtiles, visualpoints, polyspath, rasterpath, prefix, outfile)
}
\arguments{
\item{pbtiles}{txt file with the tiles names that you want to analizse}

\item{visualpoints}{shapefile where all the PB points are}

\item{polyspath}{Path where the shapefiles to analise}

\item{rasterpath}{Path to the raster files associates with the shapefiles}

\item{prefix}{Prefix of the Maxent shapefiles, if there is a prefix}

\item{outfile}{Path where you want to save the stats and graphs generated}
}
\value{
An rdsdata dataframe with the base info and the plots
}
\description{
Buffer out a point with an specific size and then compare the classes Pb and fals_pos to see if we can separate them in the post-process stage. 
It also creates the graphs to check this information.
A boxplot graph with all the quantiles and a graph to see how much fasle_pos can be extracted vs Pb retrieved.
}
