% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_maxent_output_based_on_errors_stats.r
\name{classify_maxent_output_based_on_error_stats}
\alias{classify_maxent_output_based_on_error_stats}
\title{Classify maxent output to binary presence absence maps}
\usage{
classify_maxent_output_based_on_error_stats(r_pred_dir, tile = "ALL",
  max_permittable_cutoff_final = max_permittable_cutoff_final,
  max_permittable_cutoff_npix_final = max_permittable_cutoff_npix_final,
  max_area, res, radius = 0.75, outp_dir)
}
\arguments{
\item{r_pred_dir}{A directory where binary .tifs predicting presence as 1 and absence as 0 can be found for multiple tiles}

\item{tile}{Character vector. Names of tile(s) to run. 'ALL will run all tiles in r_pred_dir. Default is 'ALL'}

\item{max_permittable_cutoff_final}{Cutoff-level for Maxent probability output. Typically produced by error_summaries.R}

\item{max_permittable_cutoff_npix_final}{Cutoff-level for the nr of pixels (clumpsize) of presence that a circles of radius radius
needs to contain to be maintained as presence. Typically produced by error_summaries.R}

\item{max_area}{The maximum are aof a polygon, all the polygons above this area will be filtered}

\item{res}{The resolution of the raster}

\item{radius}{The radius of the disk in which clump size is counted}

\item{outp_dir}{the directory to write output images to.}
}
\value{
empty
}
\description{
Classify maxent output to binary presence absence maps based on thresholds of
pixel-level probability, expected minimum crown size, and clump size
}
