% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryOperations.R
\name{COSoutput}
\alias{COSoutput}
\title{Intersect COS file}
\usage{
COSoutput(outpath, shppath, COSpath, prefix = NULL)
}
\arguments{
\item{outpath}{Path wwhere you want to save the resultin shapefiles}

\item{shppath}{Path where the Maxent shapefiles are}

\item{COSpath}{Full path plus name of the COS file}

\item{prefix}{Prefix of the Maxent shapefiles, if there is a prefix}
}
\value{
Several shapefies with the points that intersect with the COS file
}
\description{
Intersect the shapefile resulting from the last step of Maxent with the COS file.
It also can be applied to any intersection between to shapefiles
}
