% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph-finalstats.R
\name{Createdataset}
\alias{Createdataset}
\title{Outputdataset}
\usage{
Createdataset(shp_insp, listrasters, shp_poly_path, prefix, savepath)
}
\arguments{
\item{shp_insp}{Path to the visual inspected shapefile}

\item{listrasters}{csv file with the silt of locations(rasternames) you want to study.}

\item{prefix}{Prefix that you want to put to the plots in order to differenciate them. Default = NULL}

\item{savepath}{Path where you want to save the dataframe}
}
\value{
A dataframe with all the values of the validated shapefile. Used to create the stadistics of the model
}
\description{
Create a dataset of the visual inspected shapefile and save it.
}
