% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{GetPoints}
\alias{GetPoints}
\title{Sample points of presence points and background points and save them in a serialized file}
\usage{
GetPoints(tile_i, all_tifs, field_name, ninputs_tile, randompt, prob_tifs, Pols,
  vuln_classes, abs_samp, tile_dat)
}
\arguments{
\item{tile_i}{Passed by the algorithm, name of the tile to run}

\item{all_tifs}{Passed by the algorithm, list of all the tiles to run}

\item{field_name}{Character. The field in AOI.filename that contains the vuln_classes}

\item{ninputs_tile}{Integer. Number of inputs that we have fore each tile, including the tile, for exemple number of textures}

\item{randompt}{Boolean. if True random points will be added in the tiles that doesn't have any visual point. Default TRUE}

\item{prob_tifs}{Boolean (FALSE) if not wanted, Directory if wanted. tifs of each raster to run with 0 vaue for the areas that we don't want to sample 
and 1 fore the ones that we want to sample. Default FALSE}

\item{Pols}{Passed by the algorithm, PolygonDataframe Object with all the visual points assesed}

\item{vuln_classes}{A list of the classes you want to model
The list can contain one or more vectors characters. Each element of the vector represents a seperate vegetation class and response variable for the model
and the vector elements are synonyms used to describe that class
The fist place in each vector will be used in the output name used to store the calibrated model, so it should not contain spaces.
The other places should appear as attributes in the field 'field_name' of Pols}

\item{abs_samp}{Integer. How many 'absence' pixels should be randomly selected from eah tile to train the model. Default is 100.}

\item{tile_dat}{A dataframe with the points of the prevoius iteration, if the previous iteration is the initial one, an empy dataframe will be passed}
}
\value{
A dataframe with the points done for the tile selected
}
\description{
Sample points of presence points and background points and save them in a serialized file.
}
