% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryOperations.R
\name{Intersection}
\alias{Intersection}
\title{Point in polygon}
\usage{
Intersection(inpath, outpath, pointpath)
}
\arguments{
\item{inpath}{Path where the polygons shapefiles are}

\item{outpath}{Path where you weant to save the resulting point shapefile}

\item{pointpath}{Shapefile where all the PB points are}
}
\value{
A rdsdata dataframe called intersections
}
\description{
Saves in a data frame all the intersections between a polygon shapefile and a point shapefile.
Used for the Crown delineation method
}
