% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_off_selection.r
\name{MaxCutoff}
\alias{MaxCutoff}
\title{Maximum cutoff}
\usage{
MaxCutoff(input_dir, plots, stadisticspath, prefix, max_error, cla, r_pred_dir,
  tile = "ALL", max_area, res, outp_dir)
}
\arguments{
\item{input_dir}{Full path where the serialized objects with the threshold informtion are}

\item{plots}{Boolean. If TRUE plots will be created. Default FALSE
Plot-1: %of trees correctly detected % of false positives
Plot-2: error in the tree dectection
Plot-3: Distribution of the number of pixels for the trees correctly decteted}

\item{stadisticspath}{Full path where the plots will be saved}

\item{prefix}{Prefix that you want to put to the plots in order to differenciate them. Default = NULl}

\item{max_error}{Integer. Maximum error you want to select}

\item{cla}{Name of the attribute that you want to do the test on. Ex: 'Pb'}

\item{r_pred_dir}{Path where the rasters outputted by Maxent to binarize and poligonize are.}

\item{tile}{Path to a txtfile with the names of the tile you want to execute
If you want to run all the images of r_pred_dir, leave it empty. Default is 'ALL'}

\item{max_area}{Maximum area in cm. If the area detected on the image is above this value it will be erase}

\item{res}{Pixel resolution of the image. Ex: 0.30}

\item{outp_dir}{Path where you want to save the shapefiles of the binarized raster}
}
\value{
Polygon shapefile
}
\description{
Selects, given a maximum error, the cut-off probability and the minimum amout of pixels that a tree have to have. 
This information is important to do the final polygonization of the raster images outputted by MaxEnt.
After that it binarize the rasters and transformates then into a polygon shapefile.
}
