% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryOperations.R
\name{NDVI}
\alias{NDVI}
\title{NDVI by moving window}
\usage{
NDVI(rast, sizemv)
}
\arguments{
\item{rast}{raster to calculate the ndvi}

\item{sizemv}{Size desired for the moving window}
}
\value{
the resulting NDVI raster
}
\description{
Calculates the NDVI given a moving window size.
}
