% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryOperations.R
\name{ortho2012}
\alias{ortho2012}
\title{Filter shapefile with th ADS 2012 evolution on NDVI}
\usage{
ortho2012(filtpath, shppath, outpath, prefix)
}
\arguments{
\item{filtpath}{path tho the shapefiles that will act as filter}

\item{shppath}{shapefile that you want to filter}

\item{outpath}{Path where you want to save the shp generated}

\item{prefix}{Prefix of the Maxent shapefiles, if there is a prefix}
}
\value{
a shapefile filtered
}
\description{
Filter shapefile with th ADS 2012 evolution on NDVI.
}
