% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Par}
\alias{Par}
\title{Parallel enviroment}
\usage{
Par(nWorkers, data_outp_dir, data_outp_name, parallel)
}
\arguments{
\item{nWorkers}{Number of threads you want to use the maximum is Maxcores-1}

\item{data_outp_dir}{Path where you want to save the logfile}

\item{data_outp_name}{Name of the log file}

\item{parallel}{Boolean. True for parallel performance}
}
\value{
The registration of parallel flag
}
\description{
Set up the parallel enviroment if needed.
}
