% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poligonize.R
\name{Poligonize}
\alias{Poligonize}
\title{Poligonize Raster}
\usage{
Poligonize(rast, rastpath)
}
\arguments{
\item{rast}{Path to the raster to polyginze}
}
\value{
A SpatialDataFrame object
}
\description{
Poligonize a raster with values between 0 and 1.
}
