% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_off_selection.r
\name{ReadThresh}
\alias{ReadThresh}
\title{Read Threshold}
\usage{
ReadThresh(input_dir)
}
\arguments{
\item{input_dir}{Full path where the serialized objects with the threshold informtion are \link[CanHeMonR.MaxEnt]{sick_tree_errors.r}}
}
\value{
A dataframe contaning all the infor of the different threshold for the selected tiles
}
\description{
Read the thresholds calculated in: and create a dataframe with the information.
}
