% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometryOperations.R
\name{Roadoutput}
\alias{Roadoutput}
\title{Intersect road CODfile}
\usage{
Roadoutput(outpath, shppath, RoadCOSpath, prefix = NULL)
}
\arguments{
\item{outpath}{Path wwhere you want to save the resultin shapefiles.}

\item{shppath}{Path where the Maxent shapefiles are}

\item{RoadCOSpath}{Full path plus name of the road file from OSM}

\item{prefix}{Prefix of the Maxent shapefiles, if there is a prefix}
}
\value{
Several shapefies with the points that don't intersect with the Road file
}
\description{
Intersect the shapefile with the COS and roads file to erase the points that intersects with the polygons.
}
