% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.r
\name{Run_model}
\alias{Run_model}
\title{Run a saved MaxEnt model in predictive mode on a tile of image data}
\usage{
Run_model(predictor_dir, text_train_dir, MaxEntmodel_dir, fname_MaxEntmodel_r,
  output_dir, rastername, model_type, EOS)
}
\arguments{
\item{predictor_dir}{Path where predictor layers are held, rasters. If EOS = FALSE predictor_dir is a path
if EOS = True predictor_dir is the path plus the image to predict}

\item{text_train_dir}{Path where .tifs of the textures associated with r_train_dir. It is really important to avoid
errors on the execution to pass the same numer of textures per tile as in the MaxEnt model used}

\item{MaxEntmodel_dir}{Path where the MaxEnt model file is held}

\item{fname_MaxEntmodel_r}{Filename of the MaxEnt model saved in rdsdata format}

\item{output_dir}{Path to write the output to}

\item{rastername}{Character. Prefix to give the outputed raster image, for control versions}

\item{model_type}{Character. Type of model of maxent you want to use: raw, logistic or cloglog}

\item{EOS}{If EOS true the for loop will be avoided if False will work with a for loop. Default FALSE}
}
\value{
A raster image for each tile with the probabilities or cummulative probabilities of presence for each class
}
\description{
Run a saved MaxEnt model on a the image data selected.
}
\examples{
\dontrun{
Run_model(predictor_dir = "/H03_CANHEMON/Imagery/Portugal/ADS100/ortophotos_06032017/geotif/pt616000-4404000.tif",
                            text_train_dir <-'/home/martlur/Documents/TexturesAds/',
                            MaxEntmodel_dir = "/home/martlur/Documents/Dockers/docker6EOS/",
                            fname_MaxEntmodel_r = "samp10000_Pb.rdsdata",
                            output_dir = "/DATA/Results/Rcode/OutputRunSickTree",
                            rastername = "samp1000_",
                            model_type = 'cloglog',
                            loop = FALSE)
}
}
\seealso{
Depends on: \code{\link[CanHeMonR.MaxEnt]{calibrate_model.r}}
}
