% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_points.r
\name{sample_points}
\alias{sample_points}
\title{Sample training data for image classification from multiple image tiles}
\usage{
sample_points(r_train_dir, text_train_dir, tile = "ALL", text = "ALL",
  prob_tifs = FALSE, vuln_classes, training_pol_filename, field_name,
  ninputs_tile, data_outp_dir, abs_samp = 100, parallel = F, nWorkers = 4,
  data_outp_name, randompt = TRUE, EOS = FALSE)
}
\arguments{
\item{r_train_dir}{A directory where .tifs for training can be found for multiple tiles}

\item{text_train_dir}{A directory where .tifs of the textures associated with r_train_dir}

\item{tile}{Character vector or CVS file. Names of tile(s) to run a cvs file. 'ALL' will run all tiles in r_train_dir. Default is 'ALL'}

\item{text}{Character vector or CVS file. Names of text(s) to run in a cvs file. 'ALL' will run all tiles in text_train_dir. Default is 'ALL'}

\item{prob_tifs}{Boolean (FALSE) if not wanted, Directory if wanted. tifs of each raster to run with 0 vaue for the areas that we don't want to sample 
and 1 fore the ones that we want to sample. Default FALSE}

\item{vuln_classes}{A list of the classes you want to model
The list can contain one or more vectors. Each vector represents a seperate vegetation class and response variable for the model
and the vector elements are synonyms used to describe that class
The fist place in each vector will be used in the output name used to store the calibrated model, so it should not contain spaces.
The other places should appear as attributes in the field 'field_name' of Pols}

\item{training_pol_filename}{Full path to the vector file (SpatialPointsDataFrame) of which one field contains the vuln.classes}

\item{field_name}{Character. The field in AOI.filename that contains the vuln_classes}

\item{ninputs_tile}{Integer. Number of inputs that we have fore each tile, including the tile, for exemple number of textures}

\item{data_outp_dir}{The folder where you want to save the sampled data}

\item{abs_samp}{Integer. How many 'absence' pixels should be randomly selected from eah tile to train the model. Default is 100.}

\item{parallel}{Boolean. Should the code be run in parallel using the doParallel package? Default is FALSE}

\item{nWorkers}{Integer. If running the ocde in parallel, how many workers should be used? Default is 4}

\item{data_outp_name}{Character. Name of the data to output}

\item{randompt}{Boolean. if True random points will be added in the tiles that doesn't have any visual point. Default TRUE}

\item{EOS}{if EOS true the for loop will be avoided if False will work with a for loop. Default FALSE}
}
\value{
Saves a serialize object with the list with class-specific data frames of which the first column is the presence-absence response that can be used to train distribution model.
}
\description{
For each class in .shp polygon file, Sample training data for image classification from multiple image tiles using their raster bricks as predictors
}
\examples{
\dontrun{
                              }
}
