% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{Tiles}
\alias{Tiles}
\title{Treatment of tiles}
\usage{
Tiles(r_train_dir, tile, text_train_dir, text)
}
\arguments{
\item{r_train_dir}{Path where the tiles are located}

\item{tile}{Or 'ALL' for reading all the files, or the path where the csv is located with the name of the rasters}

\item{text_train_dir}{Path where the textures are located}

\item{text}{Or 'ALL' for reading all the files, or the path where the csv is located with 
the sufix of the textures wanted}
}
\value{
The list of files to process
}
\description{
Check if you want to read all the tiles from a folder or just the files needed via csv.
}
