% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph-finalstats.R
\name{Validationplots}
\alias{Validationplots}
\title{Validation plots}
\usage{
Validationplots(pathrds, validationweb = TRUE, savepath, prefix = NULL)
}
\arguments{
\item{pathrds}{Path to the serialized object with the information of the visual inspected area}

\item{savepath}{Path where you want to save the dataframe}

\item{prefix}{Prefix that you want to put to the plots in order to differenciate them. Default = NULL}
}
\value{
A series of plots of Maxent results, true positives and False positives, Accurancy tratio, error, etc.
}
\description{
Creation of the plots related to the validation of the predictions of MaxEnt.
}
