% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poligonize.R
\name{poligonize}
\alias{poligonize}
\title{poligonize Raster}
\usage{
poligonize(rast, rastpath)
}
\arguments{
\item{rastpath}{path to the raster to polyginze}
}
\value{
The Layer object
}
\description{
Poligonize a raster with values between 0 and 1
}
